source/requirements.txt
altgraph==0.17
click==7.1.2
Flask==1.1.2
fuzzyset==0.0.19
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pilkit==2.0
Pillow==7.2.0
pkg-resources==0.0.0
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.7
python-Levenshtein==0.12.0
texttable==1.6.2
Werkzeug==1.0.1
